requirements.txt
Pillow==8.1.2
spidev==3.5
RPi.GPIO==0.7.0
numpy==1.21.4
psutil==5.8.0